backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
sqlmodel==0.0.22
psycopg[binary]==3.2.3
python-docx==0.8.11
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.25.0
pydantic==2.10.3
